webSocket/requirements.txt
jsonify
pymysql
flask_cors
flask_sqlalchemy
mysql-connector-python